MC: move getAnswerText to utils.ts
[#182525578]
